Add a volume to store DB data persistently
Create a volume "booklyn_pgdata" to store data on host persistently for
the DB container. Thus, previous data can be reused when the container
is recreated unless the volume is removed.

You can remove the volume with `docker volume prune`, or remove it when
shutting down the services with `docker compose down --volumes`.